fix: post review fixes
