+++ packages/react/.eslintrc.js
@@ -11,5 +11,6 @@ module.exports = {
         'react/jsx-no-bind': 'off',
         'react/jsx-uses-react': 'off',
         'react/react-in-jsx-scope': 'off',
-    }
+        'jsx-quotes': [2, 'prefer-double'],
+    },
 };
